+++ http_cache_core/CHANGELOG.md
@@ -1,3 +1,11 @@
+## 1.1.0
+- feat: Early skip strategy calculation when request has already cache check conditions.
+- feat: Better handling of headers.
+- feat: Clamp cache-control values and other headers.
+- fix: Regression from versions 1.0.1 with `CacheOptions#maxStale` inverted behaviour.
+- chore: Add more comments.
+- chore: Add more tests & fix some.
+
 ## 1.0.2
 - fix: Enforce `CacheKeyBuilder` headers as map of `String` values instead of `dynamic`.